Make process value number a signed int

+++ src/process_runner/example_processes.rs
@@ -6,7 +6,7 @@ use crate::process_runner::state::CellState;
 pub fn example_process(cell_state: &CellState, _neighbours: &Vec<&CellState>) -> Vec<CellUpdate> {
     vec![CellUpdate {
         target_cell: cell_state.id,
-        value: Value::NumberI(cell_state.population / 10),
+        value: Value::NumberI((cell_state.population / 10) as i32),
         action: Action::ADD,
     }]
 }
@@ -21,7 +21,7 @@ pub fn population_migration(
     }
     vec![CellUpdate {
         target_cell: cell_state.id,
-        value: Value::NumberI(movement),
+        value: Value::NumberI(movement as i32),
         action: Action::ADD,
     }]
 }
